황승민 : SelectionMenu수정 시작

+++ WinAPI_HSM_PrincessMaker2/GameEngineCore/GameEngineRender.cpp
@@ -114,7 +114,7 @@ void GameEngineRender::TextRender(float _DeltaTime)
 
 	float4 RenderPos = GetActorPlusPos() - CameraPos;
 
-	TextOutA(GameEngineWindow::GetDoubleBufferImage()->GetImageDC(), RenderPos.ix(), RenderPos.iy(), RenderText.c_str(), RenderText.size());
+	TextOutA(GameEngineWindow::GetDoubleBufferImage()->GetImageDC(), RenderPos.ix(), RenderPos.iy(), RenderText.c_str(), static_cast<int>(RenderText.size()));
 
 	return;
 }
